refactor(measures): subdivide _shared/duplication/ + re-anchor budget
22 files in _shared/duplication/ exceeded the 15-fanout limit enforced by
codebase-shape.test.ts. Split into five cohesive subdirs by responsibility:

  fingerprints/  behavioral, structural, lexical
  lsh/           minhash, lsh, jaccard
  per-function/  cluster-duplicates, buckets-to-pairs, format-duplicate-rows
  workflow/      call-dag-fingerprint, cluster-call-dags, format-call-dag-rows
  ranking/       severity-ranking

extract-functions.ts stays at the top level (entry point shared by every
subdir).

Also re-anchors MAX_RECOVERABLE_LOC 2525 → 2542 against current dev-manu.
The intervening commits between the original baseline and current dev-manu
tip grew the corpus by 17 LOC of legitimate pairs; the duplication
detection itself is unchanged. The high-severity warning budget (200) and
the per-function/workflow pair-count budgets are unchanged.
