+++ packages/measures/src/_shared/duplication/per-function/buckets-to-pairs.test.ts
@@ -1,6 +1,6 @@
 import {describe, expect, it} from 'vitest'
 import {bucketsToPairs} from './buckets-to-pairs'
-import {decodePairKey} from './lsh'
+import {decodePairKey} from '../lsh/lsh'
 
 describe('bucketsToPairs', () => {
     it('produces unordered pairs for buckets with >=2 ids', () => {
