+++ packages/measures/src/_shared/duplication/ranking/severity-ranking.test.ts
@@ -1,5 +1,5 @@
 import {describe, expect, it} from 'vitest'
-import {extractFunctionsFromSource, type FunctionRecord} from './extract-functions.ts'
+import {extractFunctionsFromSource, type FunctionRecord} from '../extract-functions.ts'
 import {
     functionLoc,
     rankSeverity,
