+++ packages/measures/src/_shared/duplication/workflow/cluster-call-dags.test.ts
@@ -1,6 +1,6 @@
 import {describe, expect, it} from 'vitest'
 import {clusterCallDags} from './cluster-call-dags'
-import {extractFunctionsFromSource, type FunctionRecord} from './extract-functions'
+import {extractFunctionsFromSource, type FunctionRecord} from '../extract-functions'
 
 function recordsFrom(relativePath: string, packageName: string, source: string): FunctionRecord[] {
     return extractFunctionsFromSource(
