+++ packages/measures/src/health/duplication/semantic-duplication.test.ts
@@ -1,7 +1,7 @@
 import {describe, expect, it} from 'vitest'
-import {clusterDuplicates} from '../../_shared/duplication/cluster-duplicates'
+import {clusterDuplicates} from '../../_shared/duplication/per-function/cluster-duplicates'
 import {extractFunctions} from '../../_shared/duplication/extract-functions'
-import {formatDuplicateRows} from '../../_shared/duplication/format-duplicate-rows'
+import {formatDuplicateRows} from '../../_shared/duplication/per-function/format-duplicate-rows'
 import {discoverPackages} from '../../_shared/discovery/discover-packages'
 import {discoverSourceFiles} from '../../_shared/discovery/function-discovery'
 import {recordHealthMetric} from '../../_shared/writers/report-writer'
