+++ packages/measures/src/health/duplication/workflow-duplication.test.ts
@@ -1,7 +1,7 @@
 import {describe, expect, it} from 'vitest'
-import {clusterCallDags} from '../../_shared/duplication/cluster-call-dags'
+import {clusterCallDags} from '../../_shared/duplication/workflow/cluster-call-dags'
 import {extractFunctions} from '../../_shared/duplication/extract-functions'
-import {formatCallDagRows} from '../../_shared/duplication/format-call-dag-rows'
+import {formatCallDagRows} from '../../_shared/duplication/workflow/format-call-dag-rows'
 import {discoverPackages} from '../../_shared/discovery/discover-packages'
 import {discoverSourceFiles} from '../../_shared/discovery/function-discovery'
 import {recordHealthMetric} from '../../_shared/writers/report-writer'
